move some helper methods from kafka admin utils into kafka admin templates
